+++ manifest/init_test.go
@@ -1,0 +1,35 @@
+package manifest
+
+import (
+	"encoding/json"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/encoding/protojson"
+	"google.golang.org/protobuf/proto"
+)
+
+func assertProtoEqual(t *testing.T, expected proto.Message, actual proto.Message) {
+	t.Helper()
+
+	if !proto.Equal(expected, actual) {
+		expectedAsJSON, err := protojson.Marshal(expected)
+		require.NoError(t, err)
+
+		actualAsJSON, err := protojson.Marshal(actual)
+		require.NoError(t, err)
+
+		expectedAsMap := map[string]interface{}{}
+		err = json.Unmarshal(expectedAsJSON, &expectedAsMap)
+		require.NoError(t, err)
+
+		actualAsMap := map[string]interface{}{}
+		err = json.Unmarshal(actualAsJSON, &actualAsMap)
+		require.NoError(t, err)
+
+		// We use equal is not equal above so we get a good diff, if the first condition failed, the second will also always
+		// fail which is what we want here
+		assert.Equal(t, expectedAsMap, actualAsMap)
+	}
+}