+++ docs/release-notes/change-log.md
@@ -62,6 +62,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### CLI
 
+* Environment variables are now accepted in manifest's `imports` list.
+
+* Environment variables are now accepted in manifest's `protobuf.importPaths` list.
+
+* Fixed relative path not resolved correctly against manifest's location in `imports` list.
+
 * Changed the output modes: `module-*` modes are gone and become the
   format for `jsonl` and `json`. This means all printed outputs are
   wrapped to provide the module name, and other metadata.